requirements.txt
aiogram==3.4.1
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

motor==3.3.2         # MongoDB async client
pymongo==4.6.1       # требуется motor

redis==5.0.1         # Redis client (синхронный, сейчас подходит идеально)
